+++ packages/generator-openapi/src/utils/services.ts
@@ -28,10 +28,15 @@ export const getSummary = (document: OpenAPI.Document) => {
   return summary && summary.length < 150 ? summary : '';
 };
 
-export const buildService = (serviceOptions: Service, document: OpenAPI.Document) => {
+export const buildService = (
+  serviceOptions: Service,
+  document: OpenAPI.Document,
+  generateMarkdown?: ({}: { service: Service; document: OpenAPI.Document; markdown: string }) => string
+) => {
   const schemaPath = path.basename(serviceOptions.path as string) || 'openapi.yml';
   const documentTags = document.tags || [];
   const serviceId = serviceOptions.id || slugify(document.info.title, { lower: true, strict: true });
+  const generatedMarkdownForService = defaultMarkdown(document, schemaPath);
   return {
     id: serviceId,
     version: document.info.version,
@@ -41,7 +46,9 @@ export const buildService = (serviceOptions: Service, document: OpenAPI.Document
     specifications: {
       openapiPath: schemaPath,
     },
-    markdown: defaultMarkdown(document, schemaPath),
+    markdown: generateMarkdown
+      ? generateMarkdown({ service: serviceOptions, document, markdown: generatedMarkdownForService })
+      : generatedMarkdownForService,
     badges: documentTags.map((tag) => ({ content: tag.name, textColor: 'blue', backgroundColor: 'blue' })),
     owners: serviceOptions.owners || [],
     setMessageOwnersToServiceOwners: serviceOptions.setMessageOwnersToServiceOwners || true,