+++ packages/generator-asyncapi/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eventcatalog/sdk
 
+## 4.4.0
+
+### Minor Changes
+
+- f74f9b3: feat(plugin): added ability to mark resources as drafts
+
+## 4.3.0
+
+### Minor Changes
+
+- ac0fc47: feat(core): openapi/asyncapi plugin now supports custom markdown for …
+
 ## 4.2.1
 
 ### Patch Changes